Rewrote Luz based on Lua syntax
This uses the official Lua (5.2) syntax specification to define a sort of state
machine over a source file. It adds a parser for the syntax that generates a
set of variable-length tokens, which trigger context-sensitive state transitions.
The decompressor uses a visitor construct to hop through states and read
bits into values. See the files in doc/ for a bit more info.

Huffman coding of certain transitions & Lempel-Ziv coding is planned, as the
current file sizes are over double the size of previous methods.